Phase 4 parameter system: review fixes, FinMind backfill, domain JSON tags
Review fixes:
- FinMindProvider.GetQuotes: return error when all symbols fail
- FinMind 5sec index: parse dates in Asia/Taipei (UTC+8)
- wrapper.py: thread-local session, thread-safe sendRequest, FUBON_MOCK_MODE
- finmind_backfill.go: RateLimiter + FetchWithRetry with retry logic
- ledger.go: atomic write via temp file + rename (crash safety)
- dashboard_api.go/report.go: handle updated_at parse errors gracefully
- plugin_control.go: fix int division (conviction averaging)
- DefaultATR: 0.0 → 0.02 (reasonable fallback for position sizing)

Domain types: add JSON tags to SimulationState, DayResult, SimulationReport,
SessionSummary, BrokerRuntimeAudit, GuardOutcome, DailyBar, BacktestWindowSummary

New files:
- cmd/fubon-dma/wrapper.py: Fubon DMA session manager
- internal/marketdata/finmind_intraday.go: Taiwan 5sec index API
- internal/marketdata/finmind_backfill.go: rate limiter + retry helper

+++ internal/orchestrator/plugin_control.go
@@ -156,7 +156,7 @@ func (CIOPortfolioExecutor) Apply(agent domain.AgentSpec, recs []domain.Recommen
 
 	for _, symbol := range symbols {
 		entry := bySymbol[symbol]
-		avgConviction := entry.conviction / entry.count
+		avgConviction := int(float64(entry.conviction) / float64(entry.count))
 		if entry.count >= 3 {
 			avgConviction = int(float64(avgConviction) * params.CrowdedConvictionMultiplier.Value)
 		}
